+++ agora_node_ext/agora_rtc_engine.h
@@ -30,6 +30,8 @@
 #include "node_metadata_observer.h"
 #include "node_napi_api.h"
 #include "windows_system_api.h"
+#include "video_source_ipc.h"
+
 /*
  * Used to declare native interface to nodejs
  */
@@ -143,7 +145,6 @@ class NodeRtcEngine : public node::ObjectWrap {
   NAPI_API(setRemoteVideoStreamType);
   NAPI_API(setRemoteDefaultVideoStreamType);
   NAPI_API(enableAudioVolumeIndication);
-  NAPI_API(startAudioRecording);
   NAPI_API(stopAudioRecording);
   NAPI_API(startAudioMixing);
   NAPI_API(stopAudioMixing);
@@ -445,8 +446,22 @@ class NodeRtcEngine : public node::ObjectWrap {
   NAPI_API(setColorEnhanceOptions);
   NAPI_API(setVideoDenoiserOptions);
   NAPI_API(startEchoTestWithConfig);
+  /*
+  * 3.6.1.4
+  */
   NAPI_API(setLocalAccessPoint);
+  NAPI_API(videoSourceSetLocalAccessPoint);
   
+  /*
+  * 3.7.0
+  */
+  NAPI_API(setScreenCaptureScenario);
+  NAPI_API(enableLocalVoicePitchCallback);
+  NAPI_API(enableWirelessAccelerate);
+  NAPI_API(enableContentInspect);
+  
+  NAPI_API(enableSpatialAudio);
+  NAPI_API(setRemoteUserSpatialAudioParams);
  public:
   Isolate* getIsolate() { return m_isolate; }
   IRtcEngine* getRtcEngine() { return m_engine; }